+++ src/main/java/com/teamherb/bookstoreback/security/config/CorsConfig.java
@@ -1,4 +1,4 @@
-package com.teamherb.bookstoreback.config.security;
+package com.teamherb.bookstoreback.security.config;
 
 import java.util.Collections;
 import org.springframework.context.annotation.Bean;
