+++ src/test/resources/application.yml
@@ -1,0 +1,26 @@
+app:
+  auth:
+    token-secret: test
+    token-expiration-msec: 9999999
+
+spring:
+  profiles:
+    active: test
+  h2:
+    console:
+      enabled: true
+  datasource:
+    url: jdbc:h2:mem:bookstoredb
+    driver-class-name: org.h2.Driver
+    username: sa
+    password:
+  jpa:
+    hibernate:
+      ddl-auto: create
+    properties:
+      hibernate:
+        format_sql: true
+        use_sql_comments: true
+
+logging.level:
+  org.hibernate.SQL: debug
